style(.PopoverContent): update `btn-neutral` to context menu item style
Refactor the styling of the dialog with the `.PopoverContent` ("SHARED LINK TO CHAT" popup) class, specifically the `.btn-neutral` (when click on 3dots there, that menu), to align with the context menu item style. This includes changes to colors, hover state style (colors, remove translateY and opacity set to 1), and other relevant design elements.